CoordinateCalculator가 Point를 Drawable로 변경하도록 수정함
