+++ CoordinateCalculator/CoordinateCalculator/OutputView.swift
@@ -39,19 +39,6 @@ struct OutputView {
         }
     }
     
-    static func shaped(points: [Point]) throws -> Drawable {
-        switch points.count {
-        case 1:
-            return points[0]
-        case 2:
-            return Line(start: points[0], end: points[1])
-        case 3:
-            return Triangle(pointA: points[0], pointB: points[1], pointC: points[2])
-        default:
-            throw OutputError.shapeNotSupported
-        }
-    }
-    
     static func printShape(_ shape: Drawable) throws {
         switch shape {
         case let shape as CustomStringConvertible:
@@ -72,15 +59,11 @@ struct OutputView {
 enum OutputError: Error, CustomStringConvertible {
     
     case notDrawableRange
-    case shapeNotSupported
     
     var description: String {
         switch self {
         case .notDrawableRange:
             return "그릴 수 있는 범위가 아닙니다."
-        case .shapeNotSupported:
-            return "지원되지 않는 도형입니다."
-        }
     }
     
 }
